Configuration to separate file
